+++ src/types/DiscoverState.ts
@@ -1,0 +1,8 @@
+import { DiscoveryTrigger } from "./DiscoveryTrigger";
+
+export interface DiscoverState {
+    resetResources?: boolean;
+    requestor?: DiscoveryTrigger;
+    version?: string;
+    lastUpdated?: number;
+}